crit_from_data_t2sp

This function allows to calculate the standardized cohen's d, the critical standardized cohen's d, cohen's d, critical cohen's d and the numerator of formula from which to compute the cohen's d given the mean of the two groups, the standard deviation of the means, correlation between the two variables and sample size, specifying the degrees of freedom, the confidence level of the interval and the direction of the hypothesis.

Usage

crit_from_data_t2sp(
  m1,
  m2 = NULL,
  sd1,
  sd2 = NULL,
  r12 = NULL,
  n,
  se = NULL,
  df = NULL,
  conf.level,
  hypothesis
)

Arguments

m1

a number representing the mean of condition 1 (e.g., pre scores). If m2 is NULL, m1 is the mean of differences between condition 1 and condition 2.

m2

a number representing the mean of condition 2 (e.g., post scores). Default to NULL.

sd1

a number representing the standard deviation of condition 1. If sd2 is NULL, sd1 is the standard deviation of differences between condition 1 and condition 2.

sd2

a number representing the standard deviation of condition 2. Default to NULL.

r12

a number corresponding to the correlation between the two conditions (e.g., pre-post correlation).

n

a number corresponding to the sample size.

se

standard error of the mean of differences.

df

degrees of freedom.

conf.level

confidence level of the interval.

hypothesis

a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").

@return the output returns a dz which is the critical Cohen's d standartized on the standard deviation of the differences, the dzc is the critical standardized d using the pooled standard deviation, the d is the Cohen's d (pooled standard deviation), the bc is the numerator of the formula from which the d is calculated, df are the degrees of freedom, and se is the standard error.